### 2025-11-06 #9370 @editorialbot
[PRE REVIEW]: Solshade: Terrain-aware Solar Illumination Modelling using Digital Elevation Models and Orbital Geometry
**Submitting author:** @amanchokshi (<a href="http://orcid.org/0000-0003-1130-6390">Aman Chokshi</a>)
**Repository:** https://github.com/amanchokshi/solshade
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
