### 2025-11-14 #9370 @kthyng
@mbobra I thought this would be an interesting submission for you — can you edit this?
