### 2025-12-09 #9370 @mbobra
> [@mbobra](https://github.com/mbobra) I thought this would be an interesting submission for you — can you edit this?

@kthyng Sorry for the delay, yes!
