### 2026-03-12 #9944 @amanchokshi
Hi @groklab and @mbobra , thanks for agreeing to review `solshade`! How is it going? Is there anything I can do to help?
