### 2026-04-06 #9944 @mbobra
@groklab I know you checked the box above, but I wanted to confirm: Were you successfully able to install `solshade`?
