### 2026-04-06 #9944 @groklab
> [@groklab](https://github.com/groklab) I know you checked the box above, but I wanted to confirm: Were you successfully able to install `solshade`?

(sorry was replying on wrong account) Yes I was able to, on Win 11 WSL
