### 2026-04-09 #9944 @mbobra
@amanchokshi I cannot install solshade. I opened an issue on amanchokshi/solshade#2.
