### 2026-04-09 #9944 @amanchokshi
Hey @mbobra — thanks so much for all your help so far, I really appreciate it. Completely understand about needing to step back from reviewing.

I’ll take a look at the installation issue and follow up on [amanchokshi/solshade#2](https://github.com/amanchokshi/solshade/issues/2) shortly.
