### 2026-04-10 6de0604
Really relaxed requirements to support py3.10 <=3.14. Numba doesn't support 3.14 yet. Ran tests locally against all versions
