+++ pyproject.toml
@@ -6,9 +6,13 @@ keywords = ["solar", "terrain", "GIS", "geospatial", "DEM", "illumination"]
 authors = [{ name = "Aman Chokshi", email = "aman.chokshi@mcgill.ca" }]
 license = { text = "MIT" }
 readme = "README.md"
-requires-python = ">=3.13"
+requires-python = ">=3.10,<3.14"
 classifiers = [
   "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
 ]
@@ -24,7 +28,7 @@ dependencies = [
     "rasterio>=1.4.3",
     "rich>=14.1.0",
     "rioxarray>=0.19.0",
-    "scipy>=1.16.1",
+    "scipy>=1.15.0",
     "skyfield>=1.53",
     "typer>=0.16.0",
 ]
@@ -64,7 +68,7 @@ package = true
 dev = [
     "affine>=2.4.0",
     "ipympl>=0.9.7",
-    "ipython>=9.4.0",
+    "ipython>=8.0.0",
     "jupyter>=1.1.1",
     "jupyterlab>=4.4.5",
     "mkdocs>=1.6.1",
@@ -96,7 +100,7 @@ coverage = { cmd = "pytest --cov=solshade --cov-report=term-missing", env = { PY
 all = ["fmt", "lint", "check", "coverage"]
 
 [tool.ruff]
-target-version = "py311"
+target-version = "py310"
 line-length = 128
 src = ["src"]
 
@@ -115,7 +119,7 @@ known-first-party = ["solshade"]
 [tool.pyright]
 venvPath = "."
 venv = ".venv"
-pythonVersion = "3.11"
+pythonVersion = "3.10"
 reportUnknownMemberType = false
 reportUnknownVariableType = false
 include = ["src"]
