### 2026-04-10 8e65f3d
Added variable numba versions, py3.14 should be supported
