### 2026-04-10 f897cdc
Fixed python upperbound by limiting to <3.15. We only test <=3.14 on the CI matrix
