### 2026-04-10 4c62490
Fixed subprocess-based CLI tests
