### 2026-04-10 #9944 @amanchokshi
Hi @mbobra thanks for finding the install issue. It turned out that my `pyproject.toml` was too restrictive, and the pinned version of `numba` didn't jive with py3.14. It's been fixed, `solshade` now supports py3.10-py3.14, and `v0.1.1` has been released on PyPI ✨
