+++ .github/workflows/ci.yml
@@ -55,7 +55,7 @@ jobs:
         run: uv run pytest --cov=solshade --cov-report=xml --cov-report=term-missing
 
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v6
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
           files: ./coverage.xml
