### 2026-05-21 91d9c2e
Added contribution guide, @kvenkman noticed I was missing one
