+++ README.md
@@ -40,8 +40,9 @@ Here are some essential project links:
 - [Home page and documentation](https://solshade.readthedocs.io/en/latest)
 - [Installing solshade](https://solshade.readthedocs.io/en/latest/guides/install)
 - [Getting started with solshade](https://solshade.readthedocs.io/en/latest/guides)
-- [solshade package on PyPI](https://pypi.org/project/solshade)
+- [Solshade package on PyPI](https://pypi.org/project/solshade)
 - [Issue tracker on github](https://github.com/amanchokshi/solshade/issues)
+- [Contributing to solshade](CONTRIBUTING.md)
 
 ## Example Applications
 
