### 2026-05-23 967b6c3
Made spelling consistently British and addressed comments from @groklab
