### 2026-05-23 a33b617
Explored and added a few of the citations suggested by @groklab
