+++ paper/paper.bib
@@ -55,3 +55,49 @@
   adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft07024R},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 }
+
+@article{pvlib, 
+  doi = {10.21105/joss.00884},
+  url = {https://doi.org/10.21105/joss.00884}, 
+  year = {2018}, 
+  publisher = {The Open Journal}, 
+  volume = {3}, 
+  number = {29}, 
+  pages = {884}, 
+  author = {Holmgren, William F. and Hansen, Clifford W. and Mikofski, Mark A.}, 
+  title = {pvlib python: a python package for modeling solar energy systems}, 
+  journal = {Journal of Open Source Software} 
+}
+
+@software{topocalc,
+  author       = {Havens, Scott},
+  title        = {TopoCalc: Sky view and terrain configuration factors},
+  year         = {2021},
+  url          = {https://github.com/USDA-ARS-NWRC/topocalc}
+}
+
+@article{grass,
+title = {GRASS GIS: A multi-purpose open source GIS},
+journal = {Environmental Modelling & Software},
+volume = {31},
+pages = {124-130},
+year = {2012},
+issn = {1364-8152},
+doi = {10.1016/j.envsoft.2011.11.014},
+url = {https://www.sciencedirect.com/science/article/pii/S1364815211002775},
+author = {Markus Neteler and M. Hamish Bowman and Martin Landa and Markus Metz},
+keywords = {GIS, GRASS, OSGeo, Open source, Spatial analysis, Remote sensing},
+abstract = {The GIS software sector has developed rapidly over the last ten years. Open Source GIS applications are gaining relevant market shares in academia, business, and public administration. In this paper, we illustrate the history and features of a key Open Source GIS, the Geographical Resources Analysis Support System (GRASS). GRASS has been under development for more than 28 years, has strong ties into academia, and its review mechanisms led to the integration of well tested and documented algorithms into a joint GIS suite which has been used regularly for environmental modelling. The development is community-based with developers distributed globally. Through the use of an online source code repository, mailing lists and a Wiki, users and developers communicate in order to review existing code and develop new methods. In this paper, we provide a functionality overview of the more than 400 modules available in the latest stable GRASS software release. This new release runs natively on common operating systems (MS-Windows, GNU/Linux, Mac OSX), giving basic and advanced functionality to casual and expert users. In the second part, we review selected publications with a focus on environmental modelling to illustrate the wealth of use cases for this open and free GIS.}
+}
+
+@article{saga,
+AUTHOR = {Conrad, O. and Bechtel, B. and Bock, M. and Dietrich, H. and Fischer, E. and Gerlitz, L. and Wehberg, J. and Wichmann, V. and B\"ohner, J.},
+TITLE = {System for Automated Geoscientific Analyses (SAGA) v. 2.1.4},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {8},
+YEAR = {2015},
+NUMBER = {7},
+PAGES = {1991--2007},
+URL = {https://gmd.copernicus.org/articles/8/1991/2015/},
+DOI = {10.5194/gmd-8-1991-2015}
+}
