### 2026-05-27 3e61ef4
Release v1.0.0
