### 2026-06-03 061b54a
Added joss paper shield
