+++ README.md
@@ -3,7 +3,7 @@
 [![PyPI](https://img.shields.io/pypi/v/solshade.svg?logo=pypi&logoColor=white&label=PyPI&color=2a3573)](https://pypi.org/project/solshade/)
 [![Python](https://img.shields.io/badge/Python-3.10%2B-34777e?logo=python&logoColor=white)](https://pypi.org/project/solshade/)
 [![CI](https://img.shields.io/github/actions/workflow/status/amanchokshi/solshade/ci.yml?branch=main&logo=github&color=82af74)](https://github.com/amanchokshi/solshade/actions/workflows/ci.yml)
-[![codecov](https://img.shields.io/codecov/c/github/amanchokshi/solshade/main.svg?logo=codecov&logoColor=white&label=Coverage&color=f9dc28)](https://codecov.io/gh/amanchokshi/solshade)
+[![codecov](https://img.shields.io/codecov/c/github/amanchokshi/solshade?color=f9dc28)](https://codecov.io/gh/amanchokshi/solshade)
 [![Docs](https://img.shields.io/readthedocs/solshade/latest?logo=Read%20The%20Docs&logoColor=white&color=ea7b17)](https://solshade.readthedocs.io/en/latest/)
 [![License](https://img.shields.io/github/license/amanchokshi/EMBERS?label-License&color=b72c40&logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAdCAYAAADLnm6HAAAACXBIWXMAAB2HAAAdhwGP5fFlAAAAGXRFWHRTb2Z0d2FyZQB3d3cuaW5rc2NhcGUub3Jnm+48GgAAAflJREFUSIntljFrVUEQhb95CViYCGItGLSIIIJiISadYiOIAS0stFAEK0G0DQRs8wOs7PQvqCC2kjQJaiEICoJYaZMYzCP6WWSfLs97c/deQtJ4YODu7syZs7N7l4GOUCfVj8kmu/L0ugYCM8ChZDO7IWBPzfeOCdg5qD31urqs9q1HX11Sr6nbszl1Wl3cImkdFtXpNonm1fUOidpiXZ2vErDaELimPlDHk/9ctjaX5saTz1oD1+ogb35OD4F+RXG+Ao+AoxExGxErdVWMiJWImAUmU8y3Crd+ylV5DD31fab0gzpS4/tPBbrwDd/US8CRbDwBXKxUW4ZGvmEB9ypIquZK0cj3R4B6GjiThuvJAKbSWiuU8uUVuJ99Pwae1KkuRDmfOqFuZJfluHpM/ZXGG+rhoZjaS9iGb1CBu8Dgdj6NiNcR8RZ4nuZGgDstdl/Op+5XVzK1Z7OdnMsfD/VAUwXa8vWA28BY8nkDvBwERMQLYDkN9wK3Cnbfiq8H3MiCPwEn1H3JTgKfs/WbBQJa8Y3y9/cAuJCsDj8KBLTi6wGXgYUC4gXgSoFfK77RiHinTgHngavAKeAgMAp8AV6x+Q8/i4ifTazbzVeLrd6BNtj1nvC/gEbYvVes7v2GEAUCVtl8tbrge0SMbeVQcgR1vWIT6nu/DL8BDHb5/EeYsAMAAAAASUVORK5CYII=)](https://github.com/amanchokshi/solshade/blob/main/LICENSE)
 [![Black](https://img.shields.io/badge/Code%20Style-Black-222222.svg?logo=powershell&logoColor=white)](https://black.readthedocs.io)
